Improve top_processes() efficiency
This commit improves the efficiency in calculating
sequential "top 'x' number of processes of a given
metric" by resuming at the xth top proccess, and
appending the next y-x proccesses to the subsequent
report 'y'.

ie if -top 10 15 then after the first 10 processes
are reported instead of starting back with an empty
report and collectin the top 15, instead we now
append the next 5 proccess to the new report since
we already know the first 10.