Handle top set as 'None'
In the case where memory_info is instantiated
without the use of argparser to set defualt
values, validating the value of top would
throw an exception when set to 'None' since
the code assumes that the var is in the form
of an iterable.

Also cleans up some code where run_adb_shell_command()
calls were commented out.